chore: Testing the functionality of Topic::class after removing Grpc.php and Rest.php
